Split types codegen on field count
