+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/codegen/split/constants/JvmConstantsGen.java
@@ -15,12 +15,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.ballerinalang.compiler.bir.codegen.split;
+package org.wso2.ballerinalang.compiler.bir.codegen.split.constants;
 
 import org.ballerinalang.model.elements.PackageID;
 import org.objectweb.asm.MethodVisitor;
-import org.wso2.ballerinalang.compiler.bir.codegen.JvmTupleTypeConstantsGen;
-import org.wso2.ballerinalang.compiler.bir.codegen.JvmUnionTypeConstantsGen;
+import org.wso2.ballerinalang.compiler.bir.codegen.split.JvmCreateTypeGen;
 import org.wso2.ballerinalang.compiler.bir.model.BIRNode;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BTupleType;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BUnionType;
@@ -65,8 +64,8 @@ public class JvmConstantsGen {
     }
 
     public void setJvmCreateTypeGen(JvmCreateTypeGen jvmCreateTypeGen) {
-        unionTypeConstantsGen.setJvmCreateTypeGen(jvmCreateTypeGen);
-        tupleTypeConstantsGen.setJvmCreateTypeGen(jvmCreateTypeGen);
+        unionTypeConstantsGen.setJvmUnionTypeGen(jvmCreateTypeGen.getJvmUnionTypeGen());
+        tupleTypeConstantsGen.setJvmTupleTypeGen(jvmCreateTypeGen.getJvmTupleTypeGen());
     }
 
     public String getTupleConstantVar(BTupleType tupleType) {
